icw-python-online-shop/backend/requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
Flask-Cors
python-dotenv
Werkzeug
pymongo
